% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.provenance.R
\name{add.provenance}
\alias{add.provenance}
\title{Add data provenance to a list}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
add.provenance(
  x,
  name = NULL,
  title = NULL,
  user = NULL,
  date = NULL,
  text = NULL,
  markdown = TRUE
)
}
\arguments{
\item{x}{A list object}

\item{name}{A string identifying the data provenance entry (defaults to \code{title} without white space)}

\item{title}{A string titling the data provenance entry (defaults to \code{name})}

\item{user}{A string identifying the user (defaults to \code{Sys.getenv("LOGNAME")})}

\item{date}{A string for the date and time associated with the provenance (defaults to the current system date and time)}

\item{text}{A string containing the main text of the data provenance entry}

\item{markdown}{A logical argument specifying whether markdown formatting should be applied}
}
\description{
This function will add a custom data provenance entry in the markdown-based format used
    by the \code{borealis} package. While designed for use with shape data structures,
    the function should be compatible with any list (or an atomic vector that can be coerced
    by \code{\link{base::as.list}}.
}
\examples{

pleth <- add.provenance(
  plethodon,
  name="acquisition",
  title = "Loaded example dataset",
  text = "Started a **training analysis** using the `plethodon` dataset,
  which is included in the excellent
  [`geomorph` package](https://cran.r-project.org/web/packages/geomorph/)" )

str(pleth)

cat(pleth$provenance$acquisition)

}
