% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis.AIC.R
\name{adonis.aic}
\alias{adonis.aic}
\title{Generalized AIC for adonis models}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
adonis.aic(fit)
}
\description{
Generalized AIC for adonis models
}
