% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.adonis.R
\name{AIC.adonis}
\alias{AIC.adonis}
\title{Generalized AIC for adonis models}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
\method{AIC}{adonis}(fit)
}
\description{
Generalized AIC for adonis models
}
