% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.angle.R
\name{align.angle}
\alias{align.angle}
\title{Align a subset of landmarks by rotation from one point}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
align.angle(
  A,
  art.pt = NULL,
  angle.pts.1,
  angle.pts.2 = NULL,
  rot.pts = NULL,
  angle = 0,
  degrees = TRUE,
  show.plot = TRUE,
  plots.vertical = TRUE,
  compare.distances = TRUE,
  provenance = NULL,
  ...
)
}
\arguments{
\item{A}{A 3-dimensional array containing XY shape corrdinates for multiple specimens, or a list containing such as an array and data provenance.}

\item{art.pt}{A number specifying which landmark is the articulation point between the two landmark subsets.}

\item{angle.pts.1}{A vector or single value specifying the angle point of one subset.  If more that one value
is provided, the centroid of the landmarks described by the vector will be used; a single value
identifies a specific landmark to use.}

\item{angle.pts.2}{A vector or single value specifying the angle point of the second subset.
This could be the entire set of points of an articulated structure to be rotated.
If \code{angle.pts.2 = NULL}, then all points other than \code{angle.pts.1} and \code{art.pt} are used.}

\item{rot.pts}{A vector containing numbers specifying which landmarks are in the subset to be rotated.
If \code{NULL}, it is assumed that the points to be rotated are the same as those in \code{angle.pts.2}.}

\item{angle}{An optional value specifying the additional amount by which the rotation should be augmented (in radians).
It might be essential to use a negative angle if centroids from multiple points are used for angle points.  It should be
clear if this is the case, upon plotting results.}

\item{degrees}{A logical value specifying whether the additional rotation angle is expressed in degrees or radians (degrees is the default).}

\item{show.plot}{A logical factor or index number specifying whether to display a plot showing the alginment of one specimen relative to the reference.
If \code{show.plot = TRUE} (the default), then a specimen is plotted at random.
Alternatively, a specimen can be indicated as in \code{show.plot = 1}.
To suppress the visuakl output, set If \code{show.plot = FALSE}.}

\item{plots.vertical}{Two plots will be generated if \code{show.plot} and \code{compare.distances} are both
\code{TRUE}. This logical value specifyies whether to arrange them vertically (default) or (if \code{plots.vertical = FALSE}) horizontally.}

\item{compare.distances}{A logical value specifying whether to compare the distibutions of mean landmark distances, corrected for centroid size.
before and after alignment to the fixed angle.}

\item{provenance}{An object that should be retained for data provenance.}
}
\value{
Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.

Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.
}
\description{
A wrapper for \code{\link[geomorph]{fixed.angle}}, which rotates a subset of landmarks so that the articulation angle between subsets is constant. (See \code{?geomorph::fixed.angle})
This function adds data provenance and optional vizualization.
}
\details{
Joints can introduce nuisance variation in landmark-based geometric morphometrics.
This function rotates a subset of landmarks about a pivot point.
This step should be run before Procrustes alignment, and it is robust to differences
in the relative position, orientation and size of specimens.

The input array should contain landmark coordinates with \emph{p}, 2, and \emph{n} dimensions,
where \emph{n} is the number of specimens and \emph{p} is the number of landmarks).
Dimension 2 must have two columns that contain X and Y locations of the landmark.

If \code{show.plot = TRUE}, then landmarks from each specimen will be displayed.
Gray points indicate main structure (body) landmarks, which are not moved.
The original location of subsructure (limb / jaw) landmarks are shown with open red circles;
the adjusted location of these landmarks are shown with black points.
The function will pass extra arguments to \code{\link[borealis]{landmark.plot}},
including \code{links}. Plotting the landmarks can be useful to ensure that the reference
specimen and pivot point result in sensible alignments for the other specimens.

If the input is a \code{list}, all elements will be retained in the output.
If a \code{provenance} element is present, then it will be expanded to include an entry
for this processing step.
}
\examples{
data("plethodon", package = "geomorph")

pletho <- align.angle(
  plethodon,
  art.pt = 1,       # These three variables:
  angle.pts.1 = 6,  #   art.pt, angle.pts.1 & angle.pts.2
  angle.pts.2 = 5,  #   define the angle to align
  rot.pts = c(2:5), # rot.pts defines the points to actually rotate
  links = "chull")  # Arguments to landmark.plot are accepted too.

}
\references{
Adams and M.L. Collyer and A. Kaliontzopoulou. 2020. Geomorph: Software for geometric morphometric analyses. R package version 3.2.1. D.C. (\href{https://cran.r-project.org/package=geomorph}{Link})


}
