% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.bilat.R
\name{align.bilat}
\alias{align.bilat}
\title{Align landmarks across a line of symmetry}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
align.bilat(A, pt1, pt2, landmark.pairs, show.plot = TRUE, verbose = TRUE)
}
\arguments{
\item{A}{A 3-dimensional array containing XY shape coordinates for multiple specimens, or a list containing such as an array and data provenance.}

\item{pt1}{A number or numeric vector of landmarks specifying one of two points to define the line of symmetry.}

\item{pt2}{A number or numeric vector of landmarks specifying the second of two points to define the line of symmetry.}

\item{landmark.pairs}{A matrix with two columns. Rows pair landmarks. Landmarks in the second column
will be reflected across the line of symmetry and averaged with the paired landmark. Landmark
numbers in the first column will be retained.}

\item{show.plot}{A logical argument specifying whether to display a plot
comparing the distributions of variance in landmark distances,
corrected for centroid size, before and after alignment to the fixed angle.}

\item{verbose}{A logical argument specifying whether to display summary statistics
on the degree of asymmetry in the specimens.}
}
\value{
Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.
}
\description{
The input array should contain landmark coordinates with \emph{p}, 2, and \emph{n} dimensions,
where \emph{n} is the number of specimens and \emph{p} is the number of landmarks).
Dimension 2 must have two columns that contain X and Y locations of the landmark.
}
\examples{

This function is still under development!

}
