% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.joint.R
\name{align.joint}
\alias{align.joint}
\title{Align a subset of landmarks by rotation from one point}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
align.joint(
  A,
  substructure.LMs = NULL,
  pivot.LM = NULL,
  reference.specimen = 1,
  include.plot = TRUE,
  verbose = TRUE,
  max.iter = 10,
  provenance = NULL
)
}
\value{
Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.
}
\description{
Align a subset of landmarks by rotation from one point
}
