% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.procrustes.R
\name{align.procrustes}
\alias{align.procrustes}
\title{Generalized Procrustes alignment with interactive outlier detection and removal}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
align.procrustes(
  A,
  curves = NULL,
  show.plot.gpa = TRUE,
  outlier.analysis = FALSE,
  display.outliers = 4,
  output.gdf = TRUE,
  ...
)
}
\arguments{
\item{A}{A list or 3-dimensional array containing XY shape coordinates for multiple specimens.}

\item{curves}{An optional matrix defining which landmarks should be treated as semilandmarks by \code{gpagen}.}

\item{show.plot.gpa}{A logical value specifying whether to plot the Procrustes-aligned landmarks.}

\item{outlier.analysis}{A logical value specifying whether to perform interactive outlier analysis.
Alternatively, a numeric vector can be provided indicating the index positions of specimens to remove.
A character vector of specimen names can also be provided, indicating the specimens to remove.}

\item{display.outliers}{The number of outlier shapes to display in the interactive outlier analysis.}

\item{output.gdf}{A logical value specifying whether to return output with
a list structure that includes an element, \code{gdf}, containing a
\code{geomorph.data.frame} of shape coordinates and metadata (the default).}
}
\value{
Returns a list with Procrustes-aligned coordinates in \code{coords}, centroid size in \code{Csize}.
    All other elements from \code{\link[geomorph]{gpagen}} are in the the element \code{gpagen}.
    Any other list elements in the input are retained.
    Data provenance is also retained in \code{provenance} and updated with details of the GPA..
}
\description{
GPA is performed by a call to \code{\link[geomorph]{gpagen}}.
If \code{outlier.analysis = TRUE} then the function will display the results of
\code{\link[geomorph]{plotOutliers}}, then warp grids for the most extreme shapes.
The user is prompted to select a certain number of extreme shapes to exclude from
a repeated Procrustes alignment. This processes is iterative, until the user indicates that
it should stop or no shapes are requested for removal.
}
\examples{
data("Bombus.forewings", package = "borealis")

fw.gpa <- align.procrustes(Bombus.forewings)

fw.gpa <- align.procrustes(Bombus.forewings, show.plot.gpa = FALSE)
plot(fw.gpa$gpagen)

fw.gpa <- align.procrustes(Bombus.forewings, outlier.analysis = TRUE)
fw.gpa <- align.procrustes(Bombus.forewings, outlier.analysis = 57)
fw.gpa <- align.procrustes(Bombus.forewings,
                           outlier.analysis = c("FJ190828-002","DRA190718-001"),
                           print.progress = FALSE )

names(fw.gpa$provenance)
cat(fw.gpa$provenance$align.procrustes)

}
\references{
Adams and M.L. Collyer and A. Kaliontzopoulou. 2020. Geomorph: Software for geometric morphometric analyses. R package version 3.2.1. D.C. (\href{https://cran.r-project.org/package=geomorph}{Link})
}
