% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.reflect.R
\name{align.reflect}
\alias{align.reflect}
\title{Consistently orient specimens in an array of coordinate shape values}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
align.reflect(
  A,
  top.pt = NULL,
  bottom.pt = NULL,
  left.pt = NULL,
  right.pt = NULL,
  show.plot = TRUE,
  verbose = TRUE,
  provenance = NULL,
  ...
)
}
\arguments{
\item{A}{A 3-dimensional array containing XY shape corrdinates for multiple specimens,
or a list containing such as an array and data provenance.}

\item{top.pt}{The row number of the landmark that should be on top.}

\item{bottom.pt}{The row number of the landmark that should be on the bottom.}

\item{left.pt}{The row number of the landmark that should be on the left.}

\item{right.pt}{The row number of the landmark that should be on the right.}

\item{show.plot}{A logical factor or index number specifying whether to display a plot showing the alginment of one specimen relative to the reference.
If \code{show.plot = TRUE} (the default), then a specimen is plotted at random.
Alternatively, a specimen can be indicated as in \code{show.plot = 1}.
To suppress the visual output, set If \code{show.plot = FALSE}.}

\item{verbose}{A logical factor specifying whether to report out corrections to each specimen.}

\item{provenance}{An object that should be retained for data provenance.}
}
\value{
Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.
}
\description{
The input array should contain landmark coordinates with \emph{p}, 2, and \emph{n} dimensions,
where \emph{n} is the number of specimens and \emph{p} is the number of landmarks).
Dimension 2 must have two columns that contain X and Y locations of the landmark.
The function will pass extra arguments to \code{\link[borealis]{landmark.plot}},
including \code{links}. Plotting the landmarks can be useful to ensure that the reference
specimen and pivot point result in sensible alignments for the other specimens.
If the input is a \code{list}, all elements will be retained in the output.
If a \code{provenance} element is present, then it will be expanded to include an entry
for this processing step.
}
\details{
If no values are given for the top, bottom, left or right landmarks, then all specimens
will be oriented to match the first specimen in the array.
}
\examples{
data(plethodon, package = "geomorph")

pletho.links <- matrix(c(4,5,5,6,6,7,7,8,8,9,9,10,10,11,2,4,12,2,3,5),
                       ncol = 2, byrow = TRUE)

curated.coords <- align.reflect(
  plethodon,
  top.pt = 8, right.pt = 7,
  links = pletho.links )

}
