% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Arey301A.scope}
\alias{Arey301A.scope}
\title{Pixel-to-metric distance conversions for the dissecting scope in Arey 301A}
\format{
An object of class \code{data.frame} with 9 rows and 6 columns.
}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
Arey301A.scope
}
\description{
Ideally, micrpgraphs are taken with a ruler or other scale in the image.
    However, if that's not the case, metric distances can be infered from pixel distances
    based on this conversion table. These data are based on measurements made
    from a high-quality stage micrometer on the VWR VistaVision dissecting microscope
    in Arey 301A at Colby College. Digital photomicrographs were taken with a
    Moticam 5 digital camera. Values apply to the camera's high resolution (2592 × 1944)
    setting. However, a value for pixels per mm are also included for the
    low resolution (1296 × 972) setting. (Try to always save images at high resolution!)
}
\details{
\describe{
  \item{\code{mag}}{Total magnification}
  \item{\code{setting}}{Zoom setting}
  \item{\code{umpx}}{Microns per pixel}
  \item{\code{umpx_sd}}{Standard deviation for the micron/pixel values, based on triplicate measurements}
  \item{\code{pxmm}}{Pixels per millimeter at high resolution}
  \item{\code{pxmm_lores}}{Pixels per millimeter at low resolution}
}
}
\examples{
# Load the dataset
data("Arey301A.scope", package = "borealis")

Arey301A.scope

}
\keyword{datasets}
