% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bombus.forewings}
\alias{Bombus.forewings}
\title{Bumblebee forewing shape data}
\format{
An object of class \code{list} of length 6.
}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut]

Ye Jin "Jane" Lee \email{yjlee21@colby.edu} [cre]
}
\usage{
Bombus.forewings
}
\description{
Preliminary shape data from bumblebee forewings. These data were imported from a
    \code{tps} file using \code{\link{read.tps}} . The \code{tps} file was created from raw data using
    \code{\link{create.tps}}. These data are very preliminary. They have not been curated and have not
    undergone Procrustes alginment. The main purpose of these data are for trouble shooting
    morphometric worlkflows.
}
\details{
\describe{
  \item{\code{coords}}{An array of XY coordinates for the shape of multiple specimens (20 landmarks by 2 (XY) by 99 specimens).}
  \item{\code{tps.header}}{The tps file header, retained for provenance.}
  \item{\code{landmark.number}}{The number of landmarks in the shape data.}
  \item{\code{specimen.number}}{The number of specimens in the shape data.}
}
}
\examples{
# Load the dataset
data(Bombus.forewings)

landmark.plot(Bombus.forewings)

cat(Bombus.forewings$tps.header)

}
\keyword{datasets}
