% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bombus.tree}
\alias{Bombus.tree}
\title{A phylogeny of bumblebee species}
\format{An object of class \code{phylo} of length 8.}
\usage{
Bombus.tree
}
\description{
This phylogeny covers 26 bumblebee species (\emph{Bombus})
    focusing on those found in northeastern North America.
}
\details{
This tree is based on sequence data from five genes, reported by Cameron et al. (2007).
    The taxa in that study were subset to the species included here.
    Sequences were realigned using ClustalW and the consensus tree was inferred using
    RAxML. The \code{phylo} list also includes data on taxonomy, node support, species names,
    and convenient short specioes codes.

\describe{
  \item{\code{edge}}{A matrix encoding tree topology.}
  \item{\code{Nnode}}{The total number of nodes in the tree.}
  \item{\code{tip.label}}{A character vector with tip labels utilized by \code{phytools} plotting functions. These are species names with genus abbreviations.}
  \item{\code{edge.length}}{A numerical vector of branch lengths.}
  \item{\code{node.label}}{A character vector of clade names.}
  \item{\code{node.support}}{A character vector of bootstrap support values for each node.}
  \item{\code{code.name}}{A character vector with short codes for species names.}
  \item{\code{species.name}}{A character vector with full species names.}
}

For more information on manipulating phylogenies in R, refer to the \href{https://cran.r-project.org/web/packages/phytools/phytools.pdf}{phytools documentation}.
}
\examples{
# Load the dataset
data(Bombus.tree)

# # Plot the tree
# library("phytools")
# plot(Bombus.tree)

# # Use phytools to add extra features to the tree
# nodelabels(Bombus.tree$node.label, adj=c(1,-0.2), frame="none", cex = 0.8)

# # The tree can be copied and modified
# btree <- Bombus.tree

# # For example, the shorter code names can be used as tip labels
# btree$tip.label <- btree$code.name
# plot(btree)

# # The tree can be pruned to keep only common species
# common.species <- c("ferv","bor","tern","imp","bimac","prep","sande","vag","terri","rufo","gris")
# btree <- keep.tip(btree, common.species)
# plot(btree)

}
\references{
Cameron SA, Hines HM, and Williams PH. 2007. A comprehensive phylogeny of the bumble bees (\emph{Bombus}). \emph{Biological Journal of the Linnean Society} 91: 161-188. \url{https://doi.org/10.1111/j.1095-8312.2007.00784.x}

Revell LJ. 2011. phytools: an R package for phylogenetic comparative biology (and other things). \emph{Methods in Ecology and Evolution} 3(2): 2041-210X. \url{https://doi.org/10.1111/j.2041-210X.2011.00169.x}
}
\keyword{datasets}
