\name{borealis-package}
\alias{borealis}
\alias{borealis-package}
\docType{package}
\title{
    Bisulfite-seq OutlieR mEthylation At singLe-sIte reSolution
}
\description{
    Borealis is an R package performing outlier analysis for
    count-based bisulfite sequencing data. It detectes outlier
    methylated CpG sites from bisulfite sequencing
    (BS-seq). The core of Borealis is modeling Beta-Binomial distributions. This
    can be useful for rare disease diagnoses.
}
\details{
    See \code{packageDescription('borealis')}
}
\author{
    Maintainer: Garrett Jenkinson <gargar934@gmail.com>
}
\keyword{package}

