% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.scaled.distances.R
\name{centroid.scaled.distances}
\alias{centroid.scaled.distances}
\title{Calculate the matrix of distances between shape coordinates, scaled by centroid size}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
centroid.scaled.distances(m, scale = TRUE, symmetrical = FALSE)
}
\arguments{
\item{m}{A matrix containing XY shape corrdinates for \emph{p} landmarks}

\item{scale}{A logical argument for whether to scale the distances by centroid size}

\item{symmetrical}{A logical argument for whether to keep the output matrix symmetrical.
If \code{symmetrical = FALSE} (the default) then the upper triangle and diagonal cells
will be \code{NA}.}
}
\value{
Returns a matrix, with dimensions \emph{p} x \emph{p}.
}
\description{
Calculate the matrix of distances between shape coordinates, scaled by centroid size
}
\examples{
x <- matrix(c(9,49, 24,54, 51,28, 81,54, 81,23, 86,32), ncol = 2, byrow = TRUE)

df <- data.frame(matrix(nrow = dim(x)[1], ncol = dim(x)[1]))
for (i in (1:dim(x)[1])) {
  df[i,] <- unlist(apply(x, 1, distance, XY=x[i,]))
}
df

centroid.scaled.distances(x, symmetrical = TRUE)
centroid.scaled.distances(x, symmetrical = TRUE, scale = FALSE)

}
