% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.missing.landmarks.R
\name{estimate.missing.landmarks}
\alias{estimate.missing.landmarks}
\title{Estimate the location of missing landmarks}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
estimate.missing.landmarks(A, na.value = NA, verbose = TRUE, ...)
}
\arguments{
\item{A}{A 3-dimensional array containing XY shape coordinates for multiple specimens, or a list containing such as an array and data provenance.}

\item{na.value}{A numeric value to treat as missing data, (\code{NA}).
For example, tpsDig2 (starting with version 2.18) records missing landmarks as
coordinate values of -1.0. Setting \code{na.value = -1} will provide
compatibility with these data.}

\item{verbose}{A logical argument specifying whether to display metrics each
during each iteration.}
}
\value{
Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.
}
\description{
This function is a wrapper for \code{geomorph:estimate.missing}, which adds
records of the specimens and landmarks altered by the imputation of missing
landmarks.
}
\details{
Additional arguments are passed to \code{geomorph::estimate.missing}.
}
\examples{
j <- data("Jadera", package = "borealis")

j$coords[30,,c(1:3)] <- NA
j$coords[37,,c(1,4)] <- NA

j2 <- estimate.missing.landmarks(j)
cat(j2$provenance$estimate.missing.landmarks)
landmark.plot(j2, specimen.number = 1:4, axes = TRUE)

}
\references{
Adams and M.L. Collyer and A. Kaliontzopoulou. 2020. Geomorph: Software for geometric morphometric analyses. R package version 3.2.1. D.C. (\href{https://cran.r-project.org/package=geomorph}{Link})

Rohlf, FJ. 2015. The tps series of software. \emph{Hystrix} 26, 9–12.
(\href{https://doi.org/10.4404/hystrix-26.1-11264}{Link})
}
