% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.serial.homologs.R
\name{extract.serial.homologs}
\alias{extract.serial.homologs}
\title{Extract landmarks from serial homologs to create a new dataset}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
extract.serial.homologs(A, partitions, verbose = TRUE)
}
\arguments{
\item{A}{A 3-dimensional array containing XY shape coordinates for multiple specimens,
or a list containing such as an array and data provenance.}

\item{partitions}{A matrix or data frame listing landmarks associated with each serial
homolog. Each column should contain the landmarks for each serial homolog.}

\item{verbose}{A logical argument specifying whether to display metrics each
during each iteration.}
}
\value{
Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.
}
\description{
This function allows comparisons of shape data from serial homologs (e.g. forewing
and hindwings). Thew initial dataset must have serially homologous landmarks placed
on each serially homologous structure.
}
\examples{

}
