% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.shapes.R
\name{filter.shapes}
\alias{filter.shapes}
\title{Filter landmark coordinate shape data via a factor}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
\method{filter}{shapes}(x, index, index.description = NULL)
}
\arguments{
\item{x}{A list containing shape data in \code{coords}, \code{gpagen}, or a geomorph.data.frame in \code{gdf}, as well as (optionally) \code{metadata}.}

\item{index}{A vector for subsetting (numerical or logical). This value should correspond to metadata row numbers and specimemn numbers of the coordinate array.}

\item{index.description}{A character description of the filter criteria to be added to data provenance.}
}
\description{
The input must be a list containing one or more elements of shape data.
The function will determine the dimensionality of each element and subset them accordingly.
Matrices (2-dimensional elements) will have rows selected according to \code{index}.
Arrays (3-dimensional elements)  will have specimens (the third dimension) selected.
}
