% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Jadera}
\alias{Jadera}
\title{Shape data from \emph{Jadera haematoloma}}
\format{
An object of class \code{list} of length 6.
}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
Jadera
}
\description{
Coordinate shape data from the dorsal side of red-shouldered soapberry bugs,
    \emph{Jadera haematoloma}.
}
\details{
\describe{
  \item{\code{coords}}{An array of XY coordinates for the shape of multiple specimens (42 landmarks by 2 (XY) by 189 specimens).}
  \item{\code{landmark.number}}{The number of landmarks in the shape data.}
  \item{\code{specimen.number}}{The number of specimens in the shape data.}
  \item{\code{metadata}}{A data frame of metadata.}
  \item{\code{provenance}}{Data provenance.}
}

Metadata includes the following
\describe{
  \item{\code{id}}{An ID name for each specimen, which combines a unique identifer with the morph, sex and digitizer.}
  \item{\code{pop}}{The population of origin. \code{SP} = Silver Palm Dr, Miami, FL. \code{KL} = Key Largo, FL. \code{LE} = Leesburg, FL. \code{AC} = Aurora, CO. \code{FR} = Frederick, MD \code{PK} = Plantation Key, FL. }
  \item{\code{hostplant}}{The host plant bugs were collected from in the wild, or reared on in the lab. \code{GRT} = goldenrain tree, \emph{Koelreuteria} sp. \code{BV} = balloon vine, \emph{Cardiospermum corindum}. }
  \item{\code{seeds}}{The number of seeds caged with the cohort during the bug's rearing. (Only available for lab-raised bugs.)}
  \item{\code{cohort}}{The number of conspecifics present in the cohort during the bug's rearing. (Only available for lab-raised bugs.)}
  \item{\code{rearing}}{Whether the bug was caught in the wild or reared in the lab. }
  \item{\code{morph}}{Whether the bug was classified as long winged (LW) oir short-winged (SW). }
  \item{\code{sex}}{Whether the bug is female (f) or male (m). }
  \item{\code{digitizer}}{The initials of the person who did the landmark digitization. }
  \item{\code{bodylength}}{The length of the body, from clypeus to terminalia, measured from the ventral side (mm). }
  \item{\code{beak}}{The length of the beak (also called the rostrum or labium), from anterior of the clypeus to its distal tip (mm). }
  \item{\code{labrum}}{The length of the labrum, from anterior of the clypeus to its distal tip (mm). }
  \item{\code{antenna}}{The mean length of the antennae (mm). }
  \item{\code{femur1}}{The mean length of the femur of the first legs (mm). }
  \item{\code{femur2}}{The mean length of the femur of the second legs (mm). }
  \item{\code{femur3}}{The mean length of the femur of the third legs (mm). }
}
}
\examples{
# Load the dataset
data("Jadera", package = "borealis")

# Define connecting lines
{
  x <- 1:9
  jhae.lines <- matrix(c(x[-length(x)], x[length(x)], x[-1], x[1]), ncol = 2)
  x <- c(10,11,17,16,15,12,13,14)
  jhae.lines <- rbind(jhae.lines, matrix(c(x[-length(x)], x[length(x)], x[-1], x[1]), ncol = 2) )
  x <- c(17,18,15,16)
  jhae.lines <- rbind(jhae.lines, matrix(c(x[-length(x)], x[length(x)], x[-1], x[1]), ncol = 2) )
  x <- c(19,21,22,20,26,25,32:34,31,35:39,29,40:42)
  jhae.lines <- rbind(jhae.lines, matrix(c(x[-length(x)], x[length(x)], x[-1], x[1]), ncol = 2) )
  x <- c(19,23, 23,24, 23,28, 28,29, 19,24, 24,27, 27,28, 19,25, 27,30)
  jhae.lines <- rbind(jhae.lines, matrix(x, ncol = 2, byrow = TRUE) )
  x <- c(1,3, 7,9, 14,16, 16,18)
  jhae.lines <- rbind(jhae.lines, matrix(x, ncol = 2, byrow = TRUE) )
}

landmark.plot(Jadera, links = jhae.lines)
landmark.plot(Jadera, links = jhae.lines, specimen.number = c(23,24,31,33))

}
\references{
Fawcett, M.M., Parks, M.C., Tibbetts, A.E., Swart, J.S., Richards, E.M.,
    Vanegas, J.C., Cenzer, M., Crowley, L., Simmons, W.R., Hou, W.S., Angelini, D.R.
    (2018). Manipulation of insulin signaling phenocopies evolution of a host-associated
    polyphenism. \emph{Nature Communications} 9: 1699. \url{https://doi.org/10.1038/s41467-018-04102-1}
}
\keyword{datasets}
