% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark.plot.R
\name{landmark.plot}
\alias{landmark.plot}
\title{Plot the relative position of landmarks}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
landmark.plot(
  xy,
  square = TRUE,
  links = NULL,
  text.color = "darkred",
  line.color = "darkgray"
)
}
\arguments{
\item{xy}{A 2D matrix of X and Y corrdinates.}

\item{square}{A logical factor specifying whether the aspect ratio of the plot should be equal.}

\item{links}{A matrix with two columns indicating landmarks to connect by lines. Or \code{"chull"} can be used to draw a convex hull.}

\item{text.color}{Color names or value for the text.}

\item{line.color}{Color names or value for lines.}
}
\description{
Plot the relative position of landmarks
}
\examples{
landmark.plot(plethodon$land[,,1])

pletho.links <- matrix(c(4,5,5,6,6,7,7,8,8,9,9,10,10,11,2,4,12,4,3,5),
                       ncol = 2, byrow = TRUE)
landmark.plot(plethodon$land[,,1], links = pletho.links )

landmark.plot(plethodon$land[,,1], links = "chull")

}
