% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listed.gdf.R
\name{listed.gdf}
\alias{listed.gdf}
\title{Converts list elements gpagen and metadata into a geomorph data frame}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
listed.gdf(x)
}
\arguments{
\item{x}{A list containing the output of \code{\link[geomorph]{gpagen}}, metadata and provenance.}
}
\description{
The input must be a list containing an element of type \code{gpagen}.
This is ideally produced by \code{\link[borealis]{procrustes.alignment}}.
If the input contains a \code{metadata} element, then it and the \code{gpagen} element are
used to create a new list element of class \code{geomorph.data.frame}.
This element can be used in downstream functions. However, keeping it as an element within
a larger list object allows retension of other elements, such as data provenance.
}
\examples{
library(geomorph)
data("Bombus.forewings", package = "borealis")

names(Bombus.forewings)
cat(Bombus.forewings$provenance)

Y.gpa <- gpagen(Bombus.forewings$coords)

Y.gdf <- id.metadata.to.gdf(
  Y.gpa,
  id.factors = c("id","species","caste","digitizer"),
  provenance = Bombus.forewings$provenance
)

names(Y.gdf)

landmark.plot(Y.gdf$coords)
hist(Y.gdf$Csize)
with(Y.gdf, boxplot(Csize ~ species))
cat(Y.gdf$provenance)

}
\references{
Geomorph: Software for geometric morphometric analyses.
R package version 3.2.1. D.C. Adams and M.L. Collyer and A. Kaliontzopoulou. 2020.
(\href{https://cran.r-project.org/package=geomorph}{Link})
}
