% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient.R
\name{orient}
\alias{orient}
\title{Consistently orient specimens in an array of coordinate shape values}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
orient(
  A,
  topLM = NULL,
  bottomLM = NULL,
  leftLM = NULL,
  rightLM = NULL,
  include.plot = TRUE,
  links = NULL,
  verbose = TRUE,
  provenance = NULL
)
}
\arguments{
\item{A}{A 3-dimensional array containing XY shape corrdinates for multiple specimens,
or a list containing such as an array and data provenance.}

\item{topLM}{The row number of the landmark that should be on top.}

\item{bottomLM}{The row number of the landmark that should be on the bottom.}

\item{leftLM}{The row number of the landmark that should be on the left.}

\item{rightLM}{The row number of the landmark that should be on the right.}

\item{include.plot}{A logical factor specifying whether to include a plot showing th first specimen in the array, after orientation.}

\item{links}{A matrix with two columns indicating landmarks to connect by lines in the plot. Or \code{"chull"} can be used to draw a convex hull.}

\item{verbose}{A logical factor specifying whether to report out corrections to each specimen.}

\item{provenance}{An object that should be retained for data provenance.}
}
\value{
Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.
}
\description{
The input array should contain landmark coordinates with \emph{p}, 2, and \emph{n} dimensions,
where \emph{n} is the number of specimens and \emph{p} is the number of landmarks).
Dimension 2 must have two columns that contain X and Y locations of the landmark.
}
\details{
If no values are given for the top, bottom, left or right landmarks, then all specimens
will be oriented to match the first specimen in the array.
}
\examples{
data(plethodon, package = "geomorph")

pletho.links <- matrix(c(4,5,5,6,6,7,7,8,8,9,9,10,10,11,2,4,12,2,3,5),
                       ncol = 2, byrow = TRUE)

curated.coords <- orient(plethodon$land,
                         topLM = 8, bottomLM = 2,
                         leftLM = 12, rightLM = 7,
                         links = pletho.links )

}
