% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{pairs}
\alias{pairs}
\title{Scatterplot Matrices}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
pairs(x, col.smooth = "darkred", gap = 0.5, ...)
}
\arguments{
\item{x}{A matrix or data/frame.
Logical and factor columns are converted to numeric.
Character columns are ignored.}

\item{col.smooth}{A color to be used by \code{lines} for drawing the smooths.}

\item{gap}{A distance between subplots, in margin lines.}

\item{...}{Arguments to be passed to \code{graphics::pairs}.}
}
\description{
An enhanced version of the \code{graphics::pairs} function,
    incorporating the suggested in the example code to create upper and lower
    triangular split showing correlation coefficients and a smoothed trend line.
}
\examples{

pairs(ChickWeight)

}
