% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.alignment.R
\name{procrustes.alignment}
\alias{procrustes.alignment}
\title{A wrapper for geomorph::gpagen that includes provenance}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
procrustes.alignment(A, provenance = NULL, ...)
}
\arguments{
\item{A}{3-dimensional array containing XY shape corrdinates for multiple specimens, or a character vector of specimen IDs.}

\item{provenance}{An object that should be retained for data provenance.}
}
\value{
Returns an object of class \code{gpagen} with an added element for \code{provenance}.
}
\description{
A wrapper for geomorph::gpagen that includes provenance
}
\references{
Geomorph: Software for geometric morphometric analyses.
R package version 3.2.1. D.C. Adams and M.L. Collyer and A. Kaliontzopoulou. 2020.
(\href{https://cran.r-project.org/package=geomorph}{Link})

Rohlf, FJ. 2015. The tps series of software. \emph{Hystrix} 26, 9–12.
(\href{https://doi.org/10.4404/hystrix-26.1-11264}{Link})
}
