% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.distance.R
\name{procrustes.distance}
\alias{procrustes.distance}
\title{Procrustes distance between shapes}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
procrustes.distance(s1, s2)
}
\arguments{
\item{s1}{A 2-D matrix of shape coordinates}

\item{s2}{A 2-D matrix of shape coordinates}
}
\value{
Returns the Procrustes distance
}
\description{
Procrustes distance between shapes
}
\examples{

procrustes.distance(
  Bombus.forewings$coords[,,1],
  Bombus.forewings$coords[,,2]
)


}
