% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.jackknife.R
\name{procrustes.jackknife}
\alias{procrustes.jackknife}
\title{Examine the influence of each landmark in Procrustes alignment using jackknife re-sampling}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
procrustes.jackknife(
  A,
  curves = NULL,
  show.plot = TRUE,
  links = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{A}{A list or 3-dimensional array containing XY shape coordinates for multiple specimens.}

\item{curves}{An optional matrix defining which landmarks should be treated as semilandmarks by \code{gpagen}.}

\item{show.plot}{A logical argument specifying whether to display a plot
of the median Procrustes distances, dropping each landmark.}

\item{links}{A matrix with two columns indicating landmarks to connect by lines.}

\item{verbose}{A logical argument specifying whether to display the function's progress.}
}
\value{
The median, 0.05 and 0.95 quantiles are returned as a data frame.
    Relatively lower values will be indicative of landmarks vary more than others.
    Exercise caution in the interpretation of these results.
    Variation may be biologically meaningful or may suggest systematic error in the placement of
    the landmarks.
}
\description{
The function will remove each landmark in turn and conduct generalized Procrustes analysis, then
determine the distribution of pairwise Procrustes distances between all shapes.
}
\examples{

procrustes.jackknife(Bombus.forewings)
procrustes.jackknife(Bombus.forewings, links = "chull")

}
