% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mmm.R
\name{read.mmm}
\alias{read.mmm}
\title{Read in and convert linear multivariate morphometric data}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
read.mmm(
  input.filename = NULL,
  output.filename = NULL,
  measurement.names = NULL,
  metadata.cols = NULL,
  measurement.col = NULL,
  apply.scale = FALSE,
  invert.scale = FALSE
)
}
\arguments{
\item{input.filename}{The file name to import.}

\item{output.filename}{The file name to export. Or if \code{output.filename = TRUE},
then it will be \code{input.filename.YYMMDD.csv}, where \code{YYMMDD} is the date.}

\item{metadata.cols}{Metadata column names or numbers to be retained in the output table.}

\item{measurement.col}{A string naming the column containing measurements.
The function can make an educated guess at this, recognizing names like "Measurement", "distance" or "pixels".}

\item{apply.scale}{A string naming the column containing scale values to multiple measurement values.}

\item{invert.scale}{A logical value indicating whether to invert the scale value before applying it.}
}
\value{
Returns a list with several elements.
    \code{x} is a data frame with the measurements and metadata, each row correwsaponding to a specimen.
    \code{measurement.number} and \code{specimen.number} containing integer values.
    \code{scaled} is a logical flag for whether scaling has been applied to measurement values.
    If any specimens were missing a scale value, the the element \code{specimens.missing.scale}
    will be an ID-named index of those specimens. (These may want to be removed!)
    The element \code{provenance} is a list recording data provenance.
    If an \code{output.filename} is specified, then the data frame in \code{x} will be written to the
    file in \code{csv} or \code{tsv} format.
}
\description{
Reads in linear multivariate morphometric (MMM) data from a \code{csv} or \code{xlsx} file.
(Only the first sheet will be used from \code{xlsx} input files.)
The input spread sheet is assumed to be "long," in the sense that one column includes a list measurements
made either by hand or in grpahics software such as \href{https://imagej.net}{ImageJ}.
This organization is typically a convenient for rapid data entry.
The function reformats the multiple measurements into a traditional tabular format,
with each measurement in a seperate column.
}
\details{
There must be a column giving specimen IDs, using a name like "ID" or "specimen_IDs".
Any other columns are optional and may be used to encode metadata.
Columns named by \code{metadata.cols} will be retained, others will be ignored.
All columns in the input file will be retained if \code{metadata.cols = "all"} .

If a column name is supplied to \code{apply.scale}, then it will be used to adjust measurement values.
This is done by multiplyng the scaling factor by the measurement values.
Typically this is appropriate when scale is recorded as unit distance (e.g. mm) per pixel.
However, if scale is recorded in pixels per unit distance (e.g. pixels/mm)
it will be appropriate to set \code{invert.scale = TRUE}, if which case, the scaling factor will be
inverted before applyng it. (In other words, the scale value will be used to divide the measurement value.)
This is anologous to the treatment of scale in \code{\link[geomorph]{readland.tps}},
which applies scale values by multiplication, and in \code{\link[borelais]{create.tps}}, which allows
the user to specify how to apply the scaling factor.

Each specimen should appear with a  consequtive block of rows, with measurements in the same order.
Each measurements must appear in a consistent order for all specimens.
The number of measurements must be consistent for all specimens. Specimen metadata must appear on the
first row for each specimen. (That is, on the row for measurement 1.)

By default, the number of specimens and measurements will be determined by the function.
This will be done by using the number of cells in the specimen ID column with non-whitespace characters.
It will be assumed that all other metadata appears in the same rows, and that
any information in other rows will be ignored.
}
\examples{
mmm.data <- read.mmm(
  input.filename = "raw.measurements.csv",
  output.filename = TRUE,
  metadata.cols = "all",
  apply.scale = TRUE,
  invert.scale = TRUE
)

names(mmm.data)

dim(mmm.data$x)
head(mmm.data$x)

mmm.data$scaled

mmm.data$missing.scale

cat(unlist(mmm.data$provenance))

}
