% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.anova.table.R
\name{rmd.anova.table}
\alias{rmd.anova.table}
\title{Format anova output for R markdown}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
rmd.anova.table(m1, m2 = NULL, digits = 3, ...)
}
\arguments{
\item{m1}{A model}

\item{m2}{An optional second model. If provided, then the first model should be a null.}

\item{digits}{The number of significant digits to report in the table.
Note that this argument differs from the \code{digits} argument in \code{kable},
in that it is applied to significant digits (using \code{signif}).}
}
\description{
This function formats the table generated by \code{anova} for R markdown knitting.
}
\details{
Other arguments are passed to \code{knitr::kable}.
}
\examples{
data("mantis", package = "borealis")

x <- align.angle(mantis,
                 art.pt = 11,
                 angle.pts.1 = 1:10,
                 angle.pts.2 = 12:15,
                 rot.pts = 12:16,
                 show.plot = FALSE)

mantis.gpa <- align.procrustes(x, show.plot.gpa = FALSE)
mantis.gpa <- listed.gdf(mantis.gpa)

mantis.cs.model <- procD.lm(coords ~ log10(Csize), data = mantis.gpa$gdf, print.progress = FALSE)
mantis.model <- procD.lm(coords ~ log10(Csize) + sex, data = mantis.gpa$gdf, print.progress = FALSE)

rmd.anova.table(mantis.model)

rmd.anova.table(mantis.cs.model, mantis.model)

}
