% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree.plot.R
\name{scree.plot}
\alias{scree.plot}
\title{Plot the proportion of variance captured by each component axis of a gm.prcomp object}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
scree.plot(
  x,
  npcs = NULL,
  threshold = 0.05,
  threshold.color = "darkred",
  method = "percent",
  type = "barplot",
  ...
)
}
\arguments{
\item{x}{An object of class \code{gm.prcomp}.}

\item{npcs}{The number of components to be plotted.}

\item{threshold}{An optional threshold to display. Default is 0.05.}

\item{threshold.color}{Color for the threshold line.}

\item{method}{A string indicating how to represent the proportion of variance.
Allowable values are \code{"variance"}, \code{"proportion"}, and \code{"percentage"}. May be abbreviated.}

\item{type}{A string indicating the type of plot. Options are \code{"barplot"} and \code{"lines"}. May be abbreviated.}
}
\description{
Plot the proportion of variance captured by each component axis of a gm.prcomp object
}
\examples{
library(geomorph)
data(plethodon)

# GPA
Y.gpa <- gpagen(plethodon$land)
# Create a gm.prcomp object
PCA <- gm.prcomp(Y.gpa$coords)

# scree plot
scree.plot(PCA)

}
\references{
Geomorph: Software for geometric morphometric analyses.
R package version 3.2.1. D.C. Adams and M.L. Collyer and A. Kaliontzopoulou. 2020.
(\href{https://cran.r-project.org/package=geomorph}{Link})
}
