% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.gmm.R, R/subset.gmm.R
\name{subset.gmm}
\alias{subset.gmm}
\title{Combine shape datasets}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]

Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
\method{subset}{gmm}(A, specimens = NULL, landmarks = NULL)

\method{subset}{gmm}(A, specimens = NULL, landmarks = NULL)
}
\arguments{
\item{A}{A \code{geomorph.data.frame} or a data structure produced by functions in the \code{borealis} package.
A 3-dimensional array of XY coordinates for multiple specimens must reside in a list
element called \code{coords}, \code{gpagen$coords} or \code{gdf$coords}. Associated metadata may exist in
a matrix element called \code{metadata} or as vectors with a length equal to the number
of specimens.}

\item{specimens}{A numerical or logical vector for subsetting specimens and metadata.}

\item{landmarks}{A numerical vector for subsetting landmarks.
Subsetting landmarks will not affect metadata.}

\item{A1}{A \code{geomorph.data.frame} or a data structure produced by functions in the \code{borealis} package.
A 3-dimensional array of XY coordinates for multiple specimens must reside in a list
element called \code{coords}, \code{gpagen$coords} or \code{gdf$coords}. Associated metadata may exist in
a matrix element called \code{metadata} or as vectors with a length equal to the number
of specimens.}

\item{A2}{A second shape dataset, to be combined with \code{A1}.
The two datasets must have the same number of landmarks to allow for downstream
analyses.}

\item{group.names}{An optional character vector describing the}
}
\description{
Combine shape datasets

Subset a geometric morphometric shape dataset
}
\examples{
data(plethodon, package = "geomorph")
x <- subset.gmm(plethodon, landmarks = 1:9, specimens = 1:3)
landmark.plot(x, specimen.number = 1:4)
x$site

data(plethodon, package = "geomorph")
x <- subset.gmm(plethodon, landmarks = 1:9, specimens = 1:3)
landmark.plot(x, specimen.number = 1:4)
x$site

}
