% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGMM.R
\name{subsetGMM}
\alias{subsetGMM}
\title{Subset a geometric morphometric shape dataset}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
subsetGMM(A, specimens = NULL, landmarks = NULL)
}
\arguments{
\item{A}{A \code{geomorph.data.frame} or a data structure such as those produced by the \code{borealis} package.
A 3-dimensional array of XY coordinates for multiple specimens must reside in a list
element called \code{coords} or \code{gdf$coords}. Associated metadata may exist in
a matrix element called \code{metadata} or as vectors with a length equal to the number
of specimens.}

\item{specimens}{A vector for subsetting specimens and metadata.}

\item{landmarks}{A vector for subsetting landmarks.
Subsetting landmarks will not affect metadata.}
}
\description{
Subset a geometric morphometric shape dataset
}
\examples{
data(plethodon, package = "geomorph")
x <- subsetGMM(plethodon, landmarks = 1:9, specimens = 1:3)
landmark.plot(x, specimen.number = 1:4)
x$site

}
