% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.provenance.R
\name{write.provenance}
\alias{write.provenance}
\title{Write data provenance to a file}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
write.provenance(x, output.filename = NULL, title = NULL)
}
\arguments{
\item{x}{List containing an element for data provenance.}

\item{output.filename}{The file name to export. If done is supplied, defaults to
\code{data.provenance.YYMMDD.md}, where \code{YYMMDD} is the date.}
}
\description{
Data provenance from the \code{provenance} element of the input list will be
written to file in \href{https://www.markdownguide.org/getting-started/}{markdown} format.
}
\examples{

}
