% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.extension.R
\name{align.extension}
\alias{align.extension}
\title{Align a subset of landmarks by linear extension}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
align.extension(
  A,
  pts.1,
  pts.2 = NULL,
  distance = 0,
  reference.specimen = "all",
  show.plot = TRUE
)
}
\arguments{
\item{A}{A 3-dimensional array containing XY shape coordinates for multiple specimens, or a list containing such as an array and data provenance.}

\item{pts.1}{A vector or single value specifying the angle point of one subset.  If more that one value
is provided, the centroid of the landmarks described by the vector will be used; a single value
identifies a specific landmark to use.}

\item{pts.2}{A vector or single value specifying the angle point of the second subset.
This could be the entire set of points of an articulated structure to be rotated.
If \code{pts.2 = NULL}, then all points other than \code{pts.1} and \code{art.pt} are used.}

\item{distance}{An optional value specifying the additional amount by which the rotation should be augmented (in radians).
It might be essential to use a negative angle if centroids from multiple points are used for angle points.  It should be
clear if this is the case, upon plotting results.}

\item{reference.specimen}{A number or numeric vector specifying which specimens should be taken as the reference for the angle
defined by \code{pts.1}, \code{art.pt}, and \code{pts.2} or provided in \code{angle}.
The default is \code{"all"}, which uses the mean angle of all specimens.}

\item{show.plot}{A logical argument specifying whether to display a plot
comparing the distributions of variance in landmark distances,
corrected for centroid size, before and after alignment to the fixed angle.}
}
\value{
Returns a list with \code{coords}, \code{provenance}, and
   any other potential list elements from the input.
}
\description{
Nuisance variation in landmark positions can be introduced by extension of
certain structures. Fore example the extension of an insect's head relative
to the rest of its body.
This function moves a subset of landmarks in or out, such that all specimens
will align to a median distance or to a designated reference specimen(s).
This step should be run before Procrustes alignment, and it is robust to
differences in the relative position, orientation and size of specimens.
}
\details{
The input array should contain landmark coordinates with \emph{p}, 2, and
\emph{n} dimensions, where \emph{n} is the number of specimens and \emph{p}
is the number of landmarks). Dimension 2 must have two columns that contain
X and Y locations of the landmark. The function will pass extra arguments
to \code{\link[borealis]{landmark.plot}}, including \code{links}.
}
\examples{
# Load the dataset
data("Jadera", package = "borealis")

# Define connecting lines
{
  x <- 1:9
  jhae.lines <- matrix(c(x[-length(x)], x[length(x)], x[-1], x[1]), ncol = 2)
  x <- c(10,11,17,16,15,12,13,14)
  jhae.lines <- rbind(jhae.lines, matrix(c(x[-length(x)], x[length(x)], x[-1], x[1]), ncol = 2) )
  x <- c(17,18,15,16)
  jhae.lines <- rbind(jhae.lines, matrix(c(x[-length(x)], x[length(x)], x[-1], x[1]), ncol = 2) )
  x <- c(19,21,22,20,26,25,32:34,31,35:39,29,40:42)
  jhae.lines <- rbind(jhae.lines, matrix(c(x[-length(x)], x[length(x)], x[-1], x[1]), ncol = 2) )
  x <- c(19,23, 23,24, 23,28, 28,29, 19,24, 24,27, 27,28, 19,25, 27,30)
  jhae.lines <- rbind(jhae.lines, matrix(x, ncol = 2, byrow = TRUE) )
  x <- c(1,3, 7,9, 14,16, 16,18)
  jhae.lines <- rbind(jhae.lines, matrix(x, ncol = 2, byrow = TRUE) )
}

landmark.plot(Jadera, specimen.number = 1:4, axes = TRUE, links = jhae.lines)

j2 <- align.extension(Jadera, pts.1 = 1:9, pts.2 = 10:18)

landmark.plot(j2, specimen.number = 1:4, axes = TRUE, links = jhae.lines)

}
