% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.procrustes.R
\name{align.procrustes}
\alias{align.procrustes}
\title{Generalized Prorustes alignment with interactive outlier detection and removal}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
align.procrustes(
  A,
  curves = NULL,
  show.plot.gpa = TRUE,
  outlier.analysis = FALSE,
  display.outliers = 4,
  provenance = NULL,
  ...
)
}
\arguments{
\item{A}{A list or 3-dimensional array containing XY shape corrdinates for multiple specimens.}

\item{curves}{An optional matrix defining which landmarks should be treated as semilandmarks by \code{gpagen}.}

\item{show.plot.gpa}{A logical value specifying whether to plot the Procrustes-aligned landmarks.}

\item{outlier.analysis}{A logical value specifying whether to poerform interactive outlier analysis.}

\item{display.outliers}{The number of outlier shapes to display.}

\item{provenance}{An object that should be retained for data provenance.
Unnecessary if \code{A} is a list that already contains a \code{provenance} element.}
}
\value{
Returns a list with Procrustes-aligned coordinates in \code{coords}, centroid size in \code{Csize}.
    All other elements from \code{\link[geomorph]{gpagen}} are in the the element \code{gpagen}.
    Any other list elements in the input are retained.
    Data proveance is also retained in \code{provenance} and updated with details of the GPA..
}
\description{
GPA is performed by a call to \code{\link[geomorph]{gpagen}}.
If \code{outlier.analysis = TRUE} then the function will display the results of
\code{\link[geomorph]{plotOutliers}}, then warp grids for the most extreme shapes.
The user is prompted to select a cetain number of extreme shapes to exclude from
a repeated Procrustes alignment. This processes is iterative, until the user indicates that
it should stop or no shapes are requested for removal.
}
\examples{
data("Bombus.forewings", package = "borealis")

names(Bombus.forewings)
cat(Bombus.forewings$provenance[[1]])

fw.gpa <- align.procrustes(Bombus.forewings, show.plot.gpa = FALSE)
plot(fw.gpa)

fw.gpa <- align.procrustes(Bombus.forewings, outlier.analysis = TRUE)

names(fw.gpa$provenance)
cat(fw.gpa$provenance$GPA)

}
\references{
Adams and M.L. Collyer and A. Kaliontzopoulou. 2020. Geomorph: Software for geometric morphometric analyses. R package version 3.2.1. D.C. (\href{https://cran.r-project.org/package=geomorph}{Link})
}
