% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Euclidean distance between points in 2D space}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
distance(x1, y1, x2, y2)
}
\description{
Euclidean distance between points in 2D space
}
