% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{East1916}
\alias{East1916}
\title{Quantative Genetics Data from East 1916}
\format{
An object of class \code{data.frame} with 173 rows and 6 columns.
}
\source{
East, E. 1916. Studies on Size Inheritance in \emph{Nicotiana}. \emph{Genetics} 1(2): 164-176
    \url{http://www.genetics.org/content/1/2/164/}
}
\usage{
East1916
East1916.wide
}
\description{
A dataset reporting the corolla lengths of \emph{Nicotiana} flowers of different
    breeds, after hybridization, and after several generations of artificial selection
    for longer or shorter corolla lengths.
}
\details{
These data originally appeared in East (1916). This daatset is often used in genetics courses
    to present the effects of selection on quantitative traits.

    \code{East1916} organizes the data in a "tidy" long format. \code{East1916.wide} presents the data in
    a dateframe that most closely resembles the original Table 1 from East (1916).
}
\keyword{datasets}
