% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.metadata.to.gdf.R
\name{id.metadata.to.gdf}
\alias{id.metadata.to.gdf}
\title{Extract metadata from the specimen IDs in an array of coordinate shape data}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
id.metadata.to.gdf(
  A,
  id.factors = "id",
  seperator = "__",
  size.data = NULL,
  return.gdf = TRUE,
  keep.original.ids = FALSE,
  provenance = NULL
)
}
\arguments{
\item{A}{3-dimensional array containing XY shape corrdinates for multiple specimens, or a character vector of specimen IDs.}

\item{id.factors}{Metadata column names to be encoded in the specimen IDs.}

\item{seperator}{A character string to separate the terms used in the ID line.}

\item{size.data}{An optional vector with size data for each specimen. If input is pass as an object of class
\code{gpagen} containing \code{Csize}, it will automatically be included in the output \code{geomorph.data.frame}.}

\item{return.gdf}{A logical value indicating whether or not to override defaults (if shape data in
included in the input) and return only metadata as a \code{data.frame}. This parameter is
irrelevant if input is provided without shape data.}

\item{keep.original.ids}{By default, the ID information is parsed out in the same order as \code{id.factors}
However, if \code{keep.original.ids = TRUE} then the original ID names (encoding metadata) are retained.
The function attempts to do this in a metadata column (named by one of the elements in \code{id.factors})
that has a name like "ID", "specimen", "Sample", etc. For best effect, set the first \code{id.factors}
value as "id". Default is \code{keep.original.ids = FALSE}.}

\item{provenance}{An object that should be retained for data provenance.}
}
\description{
The function assumes that the ID names have been used to
embed metadata. This can be done if a \code{tps} file is generated using
\code{\link{create.tps}}. Check the header information or \code{ID} lines to determine
the metadata seperator. By default \code{seperator = "__"}, the same as that used
by \code{\link{create.tps}}.
}
\details{
The input is ideally an object produced by \code{\link[geomorph]{gpagen}}, in which case centroid size
(if present) will be added to the \code{geomorph.data.frame}.

The input data object can be a 3-dimensional array with ID names in the third element of \code{dimnames},
the output of \code{\link[geomorph]{gpagen}}, or a list generated by \code{\link{read.tps}}
(containing a \code{coords} element). It is also possible to pass in a character vector of ID names,
for example, directly from \code{dimnames}. The function will also recognize PCA objects of the class
\code{prcomp} or \code{gm.prcomp}. However, if PCA or character input is given, then the output will not
include shape coordinates. If the output includes shape data, then it is returned as
class \code{geomorph.data.frame}. Otherwise, the output is a regular \code{data.frame}.
}
\examples{
library(geomorph)
data("Bombus.forewings", package = "borealis")

names(Bombus.forewings)
cat(Bombus.forewings$provenance)

Y.gpa <- gpagen(Bombus.forewings$coords)

Y.gdf <- id.metadata.to.gdf(
  Y.gpa,
  id.factors = c("id","species","caste","digitizer"),
  provenance = Bombus.forewings$provenance
)

names(Y.gdf)

landmark.plot(Y.gdf$coords)
hist(Y.gdf$Csize)
with(Y.gdf, boxplot(Csize ~ species))
cat(Y.gdf$provenance)

}
\references{
Geomorph: Software for geometric morphometric analyses.
R package version 3.2.1. D.C. Adams and M.L. Collyer and A. Kaliontzopoulou. 2020.
(\href{https://cran.r-project.org/package=geomorph}{Link})

Rohlf, FJ. 2015. The tps series of software. \emph{Hystrix} 26, 9–12.
(\href{https://doi.org/10.4404/hystrix-26.1-11264}{Link})
}
