% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy.names.R
\name{phy.names}
\alias{phy.names}
\title{Add information to the names in a phylip alignment file}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
phy.names(
  input.phy = NULL,
  input.csv = NULL,
  output.filename = NULL,
  separator = "_",
  prepend = TRUE,
  strict.species.names = TRUE
)
}
\arguments{
\item{input.phy}{A phylip-format file containing an alignment with ID names
to update.}

\item{input.csv}{A comma-separated values file. This file must include a header
row, but the column names are not used. The first column must include the
existing ID names in the phylip alignment. The second column should contain
information to be added to the alignment ID names.}

\item{output.filename}{The file name to export.}

\item{separator}{A character string to separate the terms used in the names.
This string should not be a space, since the phylip format uses spaces to
distinguish names from sequence data.}

\item{prepend}{A logical value indicating whether or not new information should be
added to the beginning (TRUE, the default) or the end of the existing names.}

\item{strict.species.names}{A logical value indicating whether or not to
treat the the new information being added as a species name, and enforce
a strict genus-specific epithet format (e.g. \emph{Homo sapiens neanderthalensis}
would become \code{Homo_sapiens}).}
}
\description{
Requires a phylip (\code{phy}) file and a \code{csv} file with two columns.
The first column in the \code{csv} must match the existing names in the
\code{phy} alignment. The order of entries does not need to match, but each
name must be unique and the names must occur in both files. If no argument is
provided to \code{output.filename} then the output will over-write the input
in \code{input.phy}.
}
