% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.alignment.R
\name{procrustes.alignment}
\alias{procrustes.alignment}
\title{A wrapper for geomorph::gpagen that includes interactive outlier detection and removal}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
procrustes.alignment(
  A,
  show.plot.gpa = TRUE,
  outlier.analysis = FALSE,
  display.outliers = 4,
  provenance = NULL,
  ...
)
}
\arguments{
\item{A}{A list or 3-dimensional array containing XY shape corrdinates for multiple specimens.}

\item{show.plot.gpa}{A logical value specifying whether to plot the Procrustes-aligned landmarks.}

\item{outlier.analysis}{A logical value specifying whether to poerform interactive outlier analysis.}

\item{display.outliers}{The number of outlier shapes to display.}

\item{provenance}{An object that should be retained for data provenance.
Unnecessary if \code{A} is a list that already contains a \code{provenance} element.}
}
\value{
Returns a list with Procrustes-aligned coordinates in \code{gpagen} with an added element for \code{provenance}.
}
\description{
If \code{outlier.analysis = TRUE} then the function will display the results of
\code{\link[geomorph]{plotOutliers}}, then warp grids for the most extreme shapes.
The user is then prompted to select a cetain number of extreme shapes to exclude from
a repeated Procrustes alignment. This processes is iterative, until the user indicates that
it should stop or no shapes are requested for removal.
}
\examples{
data("Bombus.forewings", package = "borealis")

names(Bombus.forewings)
cat(Bombus.forewings$provenance[[1]])

fw.gpa <- procrustes.alignment(Bombus.forewings, show.plot.gpa = FALSE)
plot(fw.gpa)

fw.gpa <- procrustes.alignment(Bombus.forewings, outlier.analysis = TRUE)

names(fw.gpa$provenance)
cat(fw.gpa$provenance$GPA)

}
\references{
Geomorph: Software for geometric morphometric analyses.
R package version 3.2.1. D.C. Adams and M.L. Collyer and A. Kaliontzopoulou. 2020.
(\href{https://cran.r-project.org/package=geomorph}{Link})

Rohlf, FJ. 2015. The tps series of software. \emph{Hystrix} 26, 9–12.
(\href{https://doi.org/10.4404/hystrix-26.1-11264}{Link})
}
