% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tps.R
\name{read.tps}
\alias{read.tps}
\title{A wrapper function for geomorph::readland.tps}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
read.tps(
  file,
  specID = c("None", "ID", "imageID"),
  negNA = FALSE,
  readcurves = FALSE,
  warnmsg = TRUE,
  show.landmark.plot = TRUE,
  square = TRUE,
  links = NULL,
  text.color = "darkred",
  line.color = "darkgray",
  orient.specimens = FALSE,
  topLM = NULL,
  bottomLM = NULL,
  leftLM = NULL,
  rightLM = NULL,
  verbose = TRUE
)
}
\value{
Returns a list with 3 elements: \code{coords}, containing a 3-dimensional array
    of shape coordinates; and \code{landmark.number} and \code{specimen.number},
    containing integer values.
}
\description{
This function reads a \code{tps} file to obtain landmark coordinates, using
\code{\link[geomorph]{readland.tps}} and includes a few routine follow-up steps.
}
\details{
If \code{orient.specimens = TRUE} then the function will also re-orient specimens
using \code{\link{orient}}.
}
\examples{
shapes <- read.tps("sample.tps", show.landmark.plot = FALSE)

shapes <- read.tps("sample.tps")

shapes <- read.tps("sample.tps", links = "chull")

shapes <- read.tps("sample.tps",
                   links = "chull",
                   orient.specimens = TRUE,
                   verbose = FALSE)

shapes$landmark.number

shapes$specimen.number

}
\references{
Rohlf, FJ. 2015. The tps series of software. \emph{Hystrix} 26, 9–12.
(\href{https://doi.org/10.4404/hystrix-26.1-11264}{Link})

Geomorph: Software for geometric morphometric analyses.
R package version 3.2.1. D.C. Adams and M.L. Collyer and A. Kaliontzopoulou. 2020.
(\href{https://cran.r-project.org/package=geomorph}{Link})
}
