% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tps.R
\name{read.tps}
\alias{read.tps}
\title{Read shape data from a TPS file}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
read.tps(
  file,
  specID = "ID",
  negNA = FALSE,
  readcurves = FALSE,
  warnmsg = TRUE,
  extract.id.metadata = TRUE,
  keep.original.ids = FALSE,
  show.landmark.plot = TRUE,
  square = TRUE,
  links = NULL,
  text.color = "darkred",
  line.color = "darkgray",
  orient.specimens = FALSE,
  topLM = NULL,
  bottomLM = NULL,
  leftLM = NULL,
  rightLM = NULL,
  verbose = TRUE
)
}
\value{
Returns a list with several elements: \code{coords}, containing a 3-dimensional array
    of shape coordinates; and \code{landmark.number} and \code{specimen.number},
    containing integer values.
    \code{scaled} is a logical flag for whether scaling has been applied to shape coordinates.
    Data provenance is incluided in an element, \code{provenance},
    which is itself a list recording data manipulations made by functions in the
    \code{\href{https://github.com/aphanotus/borealis}{borealis}} package.
}
\description{
A wrapper function for \code{\link[geomorph]{readland.tps}}, which reads a \code{tps} file to
obtain landmark coordinates.
}
\details{
If \code{extract.id.metadata = TRUE} (the default) then the function will extracts
metadata embedded in the ID lines based on the separator and metadata factor names
described in the header of the \code{tps} file. If \code{orient.specimens = TRUE}
then the function will also re-orient specimens using \code{\link{orient}}.
}
\examples{
shapes <- read.tps("sample.tps", show.landmark.plot = FALSE)

shapes <- read.tps("sample.tps")

shapes <- read.tps("sample.tps", links = "chull")

shapes <- read.tps("sample.tps",
                   links = "chull",
                   orient.specimens = TRUE,
                   verbose = FALSE)

names(shapes)

shapes$landmark.number

shapes$specimen.number

cat(shapes$provenance)

}
\references{
Rohlf, FJ. 2015. The tps series of software. \emph{Hystrix} 26, 9–12.
(\href{https://doi.org/10.4404/hystrix-26.1-11264}{Link})

Geomorph: Software for geometric morphometric analyses.
R package version 3.2.1. D.C. Adams and M.L. Collyer and A. Kaliontzopoulou. 2020.
(\href{https://cran.r-project.org/package=geomorph}{Link})
}
