% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.provenance.R
\name{write.provenance}
\alias{write.provenance}
\title{Write data provenance to a file}
\source{
Dave Angelini \email{david.r.angelini@gmail.com} [aut, cre]
}
\usage{
write.provenance(x, output.filename = NULL, title = NULL)
}
\arguments{
\item{x}{A list of data provenance elements, or an object containing a list element called \code{provenance}.}

\item{output.filename}{The file name to export. If done is supplied, defaults to
\code{data.provenance.YYMMDD.md}, where \code{YYMMDD} is the date.}

\item{title}{An optional title for the data provenance report.}
}
\description{
Data provenance from the \code{provenance} element of the input list will be
written to file in \href{https://www.markdownguide.org/getting-started/}{markdown} format.
The \code{output.filename} may be specified, or by default it will be the input object name with the
addition of \code{provenance.YYMMDD.md}, where \code{YYMMDD} is the date.
}
\examples{

write.provenance(
  shapes.gpa,
  output.filename = "shapes.provenance.md",
  title = "Shapes data provenance"
)

}
